Add linux testing for PRs on Linux
This commit makes the following changes in support of the basics of
getting some simple testing in place.

- Reorganizes where .pony files live

This will allow for both application binary and tests binaries to be created

- Adds a Makefile

The Makefile is modelled on ponyup and corral but doesn't have any doc or installation
functionality as those aren't needed at this time. It doesn't try to do anything clever
with corral dependencies and will try to see if they need to be fetched each time.

- Removes .ci-dockerfiles, .ci-scripts, and .cirrus.yml

They aren't needed at this time. They were copied from ponyup and won't be needed in that form.

- Updates the README to reflect changed building instructions

- Removes _corral and _repos directories

It will fetch dependencies on demand as needed. We don't want to "vendor" them.

- Adds a "stub" unit test program

It can be filled out with tests as needed. Currently it has a single test that will always pass.
This is a placeholder for the CI that this commit puts into place.

- Updates .gitignore to reflect wider changes